chore: rename component
