+++ src/pages/dashboard/base/components/MiddleChart.vue
